requirements.txt
fastapi
uvicorn
openai
neo4j
pydantic
pytest
python-dotenv
